Fix chunk_index for big files, allow dump reqs
Added an environment variable to dump a prompt request and exit, for
easier testing.

Code for big files is executing, but I'm noticing but it produces
output files that are not correct and has incorrectly inserted bytes.

Will either disable or swap to different method of keeping the request
but re-acquiring the connection semaphore